+++ core/build.gradle.kts
@@ -44,8 +44,9 @@ dependencies {
     // Kotlin
     implementation(libs.kotlin)
 
-    // Lifecycle Ktx
+    // AndroidXDependencies
     implementation(libs.androidx.lifecycle.runtime.ktx)
+    implementation(libs.androidx.compose.saveable)
 
     // Hilt
     implementation(libs.hilt.android)
