+++ feature/src/main/java/com/terning/feature/onboarding/signin/SignInRoute.kt
@@ -24,12 +24,12 @@ import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.navigation.NavHostController
-import com.terning.core.designsystem.component.bottomsheet.SignUpBottomSheet
 import com.terning.core.designsystem.theme.TerningPointTheme
 import com.terning.core.extension.toast
 import com.terning.feature.R
 import com.terning.feature.home.navigation.navigateHome
 import com.terning.feature.onboarding.signin.component.KakaoButton
+import com.terning.feature.onboarding.signup.component.SignUpBottomSheet
 
 @Composable
 fun SignInRoute(
@@ -85,10 +85,8 @@ fun SignInScreen(
 
         if (showBottomSheet) {
             SignUpBottomSheet(
-
                 onDismiss = { showBottomSheet = false }
             )
-
         }
 
     }
