+++ gradle/libs.versions.toml
@@ -26,6 +26,7 @@ androidxComposeBom = "2024.04.01"
 androidxComposeCompiler = "1.5.14"
 androidxComposeMaterial3 = "1.2.1"
 composeNavigation = "2.8.0-beta04"
+composeSavable = "1.4.3"
 
 ## Hilt
 hilt = "2.48.1"
@@ -100,6 +101,7 @@ androidx-compose-ui-test = { group = "androidx.compose.ui", name = "ui-test-juni
 androidx-compose-ui-testManifest = { group = "androidx.compose.ui", name = "ui-test-manifest" }
 androidx-compose-navigation = { group = "androidx.navigation", name = "navigation-compose", version.ref = "composeNavigation" }
 androidx-compose-navigation-test = { group = "androidx.navigation", name = "navigation-testing", version.ref = "composeNavigation" }
+androidx-compose-saveable = {group = "androidx.compose.runtime", name = "runtime-saveable", version.ref =  "composeSavable"}
 compose-compiler-gradle-plugin = { module = "org.jetbrains.kotlin:compose-compiler-gradle-plugin", version.ref = "kotlin" }
 
 androidx-ui = { group = "androidx.compose.ui", name = "ui" }
